requirements.txt
fastapi
uvicorn
httpx
pyjwt
sentence-transformers
faiss-cpu
python-multipart
